requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses==0.6
Django==4.1.13
django-mongoengine==0.5.6
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
idna==3.7
mongoengine==0.28.2
pymongo==3.12.3
pytz==2024.1
requests==2.32.2
sqlparse==0.2.4
typing_extensions==4.12.0
urllib3==2.2.1
